Pull parseSingleValue/parseShorthand calls out of LegacyCSSPropertyParser
This patch moves entry call-site to parseSingleValue and parseShorthand
out of LegacyCSSPropertyParser. This means that for any property which
has been implemented in terms of CSSParserTokens, we don't have to
create CSSParserValues at all, and we never enter the code in the
LegacyCSSPropertyParser.cpp file (although we still fall-back to calling
into it if parsing fails, or CSS variables are used).

Note that this change means we will always try parsing properties with
the new path before checking for CSS variables. In the current code,
we'd notice variables when creating the CSSParserValueList and not call
into the main parsing logic. This is easy enough to change once the
logic for variables is also moved into parseValueStart.

BUG=499780

Review URL: https://codereview.chromium.org/1455693008

Cr-Original-Commit-Position: refs/heads/master@{#361603}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6e09539f06a73b0016f729a994bffa5b461df651