Suppress flaky text test failures
text-font-height-mismatch.html and text/emphasis.html have been failing
on many try runs for the last week or more on a range of CLs and it
seems clear that the failure is related to the text machine, not the
change being landed.

This is blocking VS 2015.

BUG=558574,521124,440500

Review URL: https://codereview.chromium.org/1472863005

Cr-Original-Commit-Position: refs/heads/master@{#361608}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 46afb890c62a5d2c41a2d85cc64ff9a047f87af0