Remove unnecessary assert in LayoutSVGRoot
In LayoutSVGRoot::mapLocalToContainer we were asserting that the
transform mode was UseTransforms, an assertion that was violated when
the SVG layer was squashed into certain other layers. The assertion
has been commented out for more than a year.

Testing various configurations of transforms and SVG with squashing,
such that the assert would hit, reveals no changes in rendered behavior.
Code inspection suggests that the assertion is unnecessary because
there is no place where SVG code actually checks the mode (although
there are comments that indicate we assume it is set). Furthermore,
SVGLayoutSupport::mapLocalToContainer explicitly sets the mode to
UseTransforms.

Given the lack of impact on behavior and the absence of failures or
bugs since the assertion was disabled, this patch removes the
assertion entirely.

R=pdr
BUG=231541,364901

Review URL: https://codereview.chromium.org/1477463003

Cr-Original-Commit-Position: refs/heads/master@{#361703}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1761ef2855918b3e54d974a7fa1a860d604d660f