Fix crash in ObjectPainter::paintOutline in flipped blocks
This patch fixes an out-of-bounds access to a Vector, when the CSS
outline property is applied to an inline element in flipped blocks
writing mode.

crbug.com/491454 fixed the positioning of the outline in the case, but
it did not properly handle when non-empty outline becomes empty by the
transform.

BUG=561053, 491454

Review URL: https://codereview.chromium.org/1481483002

Cr-Original-Commit-Position: refs/heads/master@{#361790}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5c70b66b1b9b5488460c8c048ff7c32542d4c36f