[css-flexbox] Correctly implement flex factors < 1
When the sum of flex factors is < 1, we need to multiply the available
space by that. We currently do this in a slightly different way that's
mostly-correct for flex-grow, but doesn't do anything for flex-shrink.

This patch fixes that.

BUG=480752,549993
R=leviw@chromium.org

Review URL: https://codereview.chromium.org/1470373003

Cr-Original-Commit-Position: refs/heads/master@{#361808}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36921b28f83a30d8d0caaa408b1586afed2ca323