Oilpan: MatchedPropertiesCache must be cleared promptly
Some destructors in the properties (e.g., ~FontFallbackList) expect that
the destructors are called promptly without relying on a GC timing.
Thus we need to clear the MatchedPropertiesCache when the StyleResolver
is cleared (not when the StyleResolver is destructed by a GC).

BUG=

Review URL: https://codereview.chromium.org/1477723003

Cr-Original-Commit-Position: refs/heads/master@{#361837}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f18e597b946074fbbce25dba5d602293a599811b