[PartitionAlloc] Take type argument for heap profiling
This adds an extra argument to |PartitionAlloc| and similar allocating
functions: the type name. The type name is then passed to the heap
profiling hook. A new macro, |WTF_HEAP_PROFILER_TYPE_NAME| is introduced
to obtain a type name string. On non-official builds, this macro expands
to a null pointer, so the impact on official builds is minimal. On
non-official builts, the macro expands to the type name, so developers
can get type info in the heap profiler without rebuilding.

In the future, some instances of |WTF_HEAP_PROFILER_TYPE_NAME| can be
replaced to pass a type name string unconditionally. This will enable
type info in the heap profiler even for official builds, without
bloating the binary, because only the most common types should be
annotated.

This is part of the heap profiler in chrome://tracing.

BUG=524631

Review URL: https://codereview.chromium.org/1472083008

Cr-Original-Commit-Position: refs/heads/master@{#362459}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 006dc1e53075a580f789a2b18f7a47316e217dc7