Fix UnitType bug in LengthSVGInterpolation.cpp
https://codereview.chromium.org/1421533006 converts SVGLength to use CSS
primitive types, introducing a UserUnits value for CSSPrimitiveValue::UnitType.
In the process of converting an SVGLength to an InterpolableValue, values like
"10px" are receiving a UnitType of Number, which causes a crash when applying
the SVG attribute. (See the bug for more details.)

This patch fixes the crash by using the new UnitType of UserUnits instead of
Number in the result of the conversion.

BUG=563303

Review URL: https://codereview.chromium.org/1481403004

Cr-Original-Commit-Position: refs/heads/master@{#362621}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d24815f4f2ee62f5bec4b814224ca1b5dbe3ec29