Handled neutered ImageData in texImage2D and texSubImage2D
Ideal we would like to throw an exception in this case. But since the
API for these two methods does not have RaiseException, this CL handles
this case the same way as the case when ImageData is null.

BUG=564846

Review URL: https://codereview.chromium.org/1498603002

Cr-Original-Commit-Position: refs/heads/master@{#363069}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c5c1f8652c7b99ec36753ed056cd4e480ad53865