Implement 'dirtiness' of HTMLOptionElement.
https://html.spec.whatwg.org/multipage/forms.html#concept-option-dirtiness

The flag is referred when 'selected' attribute is added or removed.

This flag is set to true when a user udpates selectedness explicitly, or
HTMLSelectElement::value setter, selectedIndex setter, or
HTMLOptionElement::selected setter.
The flag is reset to false on resetting <select>.

We need to inrorudce HTMLOptionElement::setSelectedForBinding because
setSelected is used internally, and the dirtiness should not be changed for the
internal callsites.

BUG=307348

Review URL: https://codereview.chromium.org/1493043005

Cr-Original-Commit-Position: refs/heads/master@{#363172}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a17cc1eda58ad8270dc56ebafab215e67e6c1bfe