Remove userCancelledLoad and refactor the code at the call sites
The function userCancelledLoad isn't called when the user cancels
the load (there's no way to do that) and yet it implements that
bit of the spec. So removing this function and adding only
the necessary bits at the call sites.

Review URL: https://codereview.chromium.org/1481913002

Cr-Original-Commit-Position: refs/heads/master@{#363193}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fa2a5db2eb8b92301b06f65aa644aef1a5bf709c