Remove unintentional use of DISALLOW_COPY.
This macro is part of wtf/dtoa/* library, and should not be used in
Blink code. This is the only occurrence of this macro in Blink.

This patch replaces it with WTF_MAKE_NONCOPYABLE, which is more
restrictive than DISALLOW_COPY. In this specifc case that is fine.

BUG=565932

Review URL: https://codereview.chromium.org/1502543002

Cr-Original-Commit-Position: refs/heads/master@{#363329}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5e2ea322510e39874c7db98c379bbfc58df5f0cc