Eliminate [LegacyInterfaceTypeChecking] for modules/filesystem bindings
(Part 2 of 2; follow up to https://crrev.com/1480363003)

Let bindings generate a TypeError, rather than the module code
producing a TypeMismatchError or InvalidModificationError, if null is
passed for non-nullable interface arguments. Previously untested,
so new test cases are added.

This also changes how the errors are reported. The async moveTo() and
copyTo() methods on Entry objects report failures via a callback and
`error` attribute. This continues to apply if the error is caused by
asynchronous filesystem operations, but per Web IDL invalid arguments
now will result in synchronous exception instead.

Given that this applies only in cases that would always have failed
(i.e. wrong type of argument is passed, or null), we believe this change
is safe but are landing it in a separate CL for ease of reverting.

BUG=561338
R=philipj@opera.com

Review URL: https://codereview.chromium.org/1499303002

Cr-Original-Commit-Position: refs/heads/master@{#363372}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b8e6a42d49ec47adc219ef8fb57db64ea88bb4c9