[css-grid] Get rid of GridResolvedPosition
GridResolvedPosition was a small class just wrapping a size_t.
In the future it should actually wrap an integer
(as we want to support implicit tracks before the explicit grid).

The class itself is not providing any benefit,
so we can get rid of it and store directly 2 size_t in GridSpan.

This will make simpler future changes related to this task.

We keep the class just as a utility for the methods
that deal with the positions resolution.
But it should be renamed in a follow-up patch.

No new tests, no change of behavior.

BUG=444011

Review URL: https://codereview.chromium.org/1500433003

Cr-Original-Commit-Position: refs/heads/master@{#363384}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3ad965028770aec5042ea3587843aec66027e1da