[Line Layout API] Convert AbstractInlineTextBox to new line layout API
This affects AXInlineTextBox, which previously was the only caller to
AbstractInlineTextBox::layoutText(), and is now the only caller to
AbstractInlineTextBox::lineLayoutItem().

This also affects InlineTextBox, in the sense that AbstractInlineTextBox
was the last remaining caller to InlineTextBox::layoutObject(), which
is now removed.

BUG=499321

Review URL: https://codereview.chromium.org/1433253003

Cr-Original-Commit-Position: refs/heads/master@{#363853}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 24f1ae8c7c39f5bd7e03310b91c26b4ce1f93a3f